@@ -10,22 +10,21 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "bs":"bash run.sh"
+    "bs": "bash run.sh"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.12",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.2.12",
-    "@angular/compiler": "^7.2.12",
-    "@angular/core": "^7.2.12",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
     "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "7.2.12",
-    "@angular/http": "7.2.12",
+    "@angular/forms": "8.2.14",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "7.2.12",
-    "@angular/platform-browser-dynamic": "7.2.12",
-    "@angular/router": "7.2.12",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
     "angular-pipes": "^9.0.2",
     "angular-tree-component": "^8.3.0",
     "aws-sdk": "^2.445.0",
@@ -34,11 +33,11 @@
     "compression": "^1.7.3",
     "cookie-parser": "^1.4.4",
     "cors": "^2.8.5",
-    "cron":"^1.8.2",
+    "cron": "^1.8.2",
     "dotenv": "^7.0.0",
     "events": "^3.0.0",
     "express": "^4.16.4",
-    "fs-extra":"^8.1.0",
+    "fs-extra": "^8.1.0",
     "hammerjs": "^2.0.8",
     "helmet": "^3.15.0",
     "http-errors": "^1.7.1",
@@ -50,22 +49,23 @@
     "ng2-dragula": "^2.1.1",
     "ngx-cookie-service": "^2.1.0",
     "recordrtc": "5.5.8",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.5",
     "rxjs-compat": "^6.3.3",
     "swagger-ui-express": "^4.0.2",
-    "uuid" : "^3.3.3",
+    "uuid": "^3.3.3",
     "wavesurfer.js": "^2.1.3",
-    "zone.js": "0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.8",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "7.2.12",
-    "@angular/language-service": "7.2.12",
+    "@angular-devkit/build-angular": "~0.803.25",
+    "@angular/cli": "^8.3.26",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
     "@types/jasmine": "~3.3.4",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "^10.12.18",
     "@types/wavesurfer.js": "^2.0.2",
+    "codelyzer": "^5.0.1",
     "concurrently": "^4.1.0",
     "eslint": "^5.16.0",
     "eslint-config-standard": "^12.0.0",
@@ -80,7 +80,7 @@
     "@types/dom-mediacapture-record": "^1.0.1",
     "@types/react": "^16.8.13",
     "@types/webrtc": "^0.0.24",
-    "typescript": ">=3.1.1 < 3.2"
+    "typescript": "3.5.3"
   },
   "prettier": {
     "singleQuote": true
